+++ lib/src/rive.dart
@@ -1,4 +1,8 @@
+import 'package:flutter/gestures.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+import 'package:rive/src/controllers/state_machine_controller.dart';
 import 'package:rive/src/rive_core/artboard.dart';
 import 'package:rive/src/rive_render_box.dart';
 import 'package:rive/src/runtime_artboard.dart';
@@ -38,6 +42,22 @@ class Rive extends LeafRenderObjectWidget {
   /// Enables/disables antialiasing
   final bool antialiasing;
 
+  /// Enable input listeners (such as hover, pointer down, etc.) on the
+  /// artboard.
+  ///
+  /// Default `false`.
+  final bool enablePointerEvents;
+
+  /// The mouse cursor for mouse pointers that are hovering over the region.
+  ///
+  /// When a mouse enters the region, its cursor will be changed to the [cursor].
+  /// When the mouse leaves the region, the cursor will be decided by the region
+  /// found at the new location.
+  ///
+  /// The [cursor] defaults to [MouseCursor.defer], deferring the choice of
+  /// cursor to the next region behind it in hit-test order.
+  final MouseCursor cursor;
+
   /// {@template Rive.clipRect}
   /// Clip the artboard to this rect.
   ///
@@ -51,6 +71,8 @@ class Rive extends LeafRenderObjectWidget {
     required this.artboard,
     this.useArtboardSize = false,
     this.antialiasing = true,
+    this.enablePointerEvents = false,
+    this.cursor = MouseCursor.defer,
     BoxFit? fit,
     Alignment? alignment,
     this.clipRect,
@@ -66,7 +88,9 @@ class Rive extends LeafRenderObjectWidget {
       ..alignment = alignment
       ..artboardSize = Size(artboard.width, artboard.height)
       ..useArtboardSize = useArtboardSize
-      ..clipRect = clipRect;
+      ..clipRect = clipRect
+      ..enableHitTests = enablePointerEvents
+      ..cursor = cursor;
   }
 
   @override
@@ -79,13 +103,20 @@ class Rive extends LeafRenderObjectWidget {
       ..alignment = alignment
       ..artboardSize = Size(artboard.width, artboard.height)
       ..useArtboardSize = useArtboardSize
-      ..clipRect = clipRect;
+      ..clipRect = clipRect
+      ..enableHitTests = enablePointerEvents
+      ..cursor = cursor;
   }
 }
 
-class RiveRenderObject extends RiveRenderBox {
+class RiveRenderObject extends RiveRenderBox implements MouseTrackerAnnotation {
   RuntimeArtboard _artboard;
-  RiveRenderObject(this._artboard) {
+  RiveRenderObject(
+    this._artboard, {
+    MouseCursor cursor = MouseCursor.defer,
+    bool validForMouseTracker = true,
+  })  : _cursor = cursor,
+        _validForMouseTracker = validForMouseTracker {
     _artboard.redraw.addListener(scheduleRepaint);
   }
 
@@ -99,6 +130,110 @@ class RiveRenderObject extends RiveRenderBox {
     _artboard = value as RuntimeArtboard;
     _artboard.redraw.addListener(scheduleRepaint);
     markNeedsLayout();
+  }
+
+  /// Local offset to global artboard position
+  Vec2D _toArtboard(Offset local) {
+    final globalCoordinates = localToGlobal(local);
+    return globalToArtboard(globalCoordinates);
+  }
+
+  /// Helper to manage hit testing
+  void _hitHelper(PointerEvent event,
+      void Function(StateMachineController, Vec2D) callback) {
+    final artboardPosition = _toArtboard(event.localPosition);
+    final stateMachineControllers =
+        _artboard.animationControllers.whereType<StateMachineController>();
+    for (final stateMachineController in stateMachineControllers) {
+      callback(stateMachineController, artboardPosition);
+    }
+  }
+
+  @override
+  void handleEvent(PointerEvent event, HitTestEntry entry) {
+    assert(debugHandleEvent(event, entry));
+    if (!enableHitTests) {
+      return;
+    }
+    if (event is PointerDownEvent) {
+      _hitHelper(
+        event,
+        (controller, artboardPosition) =>
+            controller.pointerDown(artboardPosition, event),
+      );
+    }
+    if (event is PointerUpEvent) {
+      _hitHelper(
+        event,
+        (controller, artboardPosition) =>
+            controller.pointerUp(artboardPosition),
+      );
+    }
+    if (event is PointerMoveEvent) {
+      _hitHelper(
+        event,
+        (controller, artboardPosition) =>
+            controller.pointerMove(artboardPosition),
+      );
+    }
+    if (event is PointerHoverEvent) {
+      _hitHelper(
+        event,
+        (controller, artboardPosition) =>
+            controller.pointerMove(artboardPosition),
+      );
+    }
+  }
+
+  @override
+  PointerEnterEventListener? get onEnter => (event) {
+        if (!enableHitTests) return;
+        _hitHelper(
+          event,
+          (controller, artboardPosition) =>
+              controller.pointerEnter(artboardPosition),
+        );
+      };
+
+  @override
+  PointerExitEventListener? get onExit => (event) {
+        if (!enableHitTests) return;
+        _hitHelper(
+          event,
+          (controller, artboardPosition) =>
+              controller.pointerExit(artboardPosition),
+        );
+      };
+
+  @override
+  MouseCursor get cursor => _cursor;
+  MouseCursor _cursor;
+  set cursor(MouseCursor value) {
+    if (_cursor != value) {
+      _cursor = value;
+      // A repaint is needed in order to trigger a device update of
+      // [MouseTracker] so that this new value can be found.
+      markNeedsPaint();
+    }
+  }
+
+  @override
+  bool get validForMouseTracker => _validForMouseTracker;
+  bool _validForMouseTracker;
+
+  @override
+  void attach(PipelineOwner owner) {
+    super.attach(owner);
+    _validForMouseTracker = true;
+  }
+
+  @override
+  void detach() {
+    // It's possible that the renderObject be detached during mouse events
+    // dispatching, set the [MouseTrackerAnnotation.validForMouseTracker] false to prevent
+    // the callbacks from being called.
+    _validForMouseTracker = false;
+    super.detach();
   }
 
   @override